+++ src/Bozboz/Admin/Reports/Downloadable.php
@@ -1,0 +1,12 @@
+<?php
+
+namespace Bozboz\Admin\Reports;
+
+interface Downloadable
+{
+	/**
+	 * @param  Bozboz\Admin\Models\Base  $instance
+	 * @return array
+	 */
+	public function getColumnsForCSV($instance);
+}