+++ src/Bozboz/MediaLibrary/Controllers/MediaLibraryAdminController.php
@@ -3,11 +3,18 @@
 use Bozboz\Admin\Controllers\ModelAdminController;
 use Bozboz\Admin\Reports\Report;
 use Bozboz\MediaLibrary\Decorators\MediaAdminDecorator;
-use View, Response, Request, Input, Redirect, Str;
+use Bozboz\MediaLibrary\Models\Media;
+use View, Response, Request, Input, Redirect, Str, File;
+
+use Symfony\Component\HttpFoundation\File\UploadedFile;
 
 class MediaLibraryAdminController extends ModelAdminController
 {
 	protected $createView = 'admin::media.upload';
+	protected $mimeTypeMapping = [
+		'image/*' => 'image',
+		'application/pdf' => 'pdf'
+	];
 
 	public function __construct(MediaAdminDecorator $media)
 	{
@@ -23,6 +30,12 @@ class MediaLibraryAdminController extends ModelAdminController
 		$report->overrideView('admin::media.overview');
 		return $report->render(array('controller' => get_class($this)));
 	}
+	
+	public function viewPrivate($id)
+	{
+		$media = Media::find($id);
+		return Response::download(storage_path().$media->getFileName());
+	}
 
 	private function ajaxJSONData()
 	{
@@ -35,6 +48,7 @@ class MediaLibraryAdminController extends ModelAdminController
 				'id' => $inst->id,
 				'caption' => $inst->caption ? $inst->caption : $inst->filename,
 				'filename' => $inst->filename,
+				'private' => $inst->private,
 				'type' => $inst->type
 			);
 		}
@@ -50,33 +64,64 @@ class MediaLibraryAdminController extends ModelAdminController
 	{
 		$data = [];
 		$captions = Input::get('caption', []);
+		$is_private = Input::get('is_private', []);
 
 		if (Input::hasFile('files')) {
 			foreach(Input::file('files') as $index => $file) {
 				$newMedia = $this->decorator->newModelInstance($file);
-				$type = explode('/', $file->getMimeType())[0];
-				$filename = $this->cleanFilename($file->getClientOriginalName());
-				$uploadSuccess = $file->move(public_path('media/' . $type), $filename);
+
+				$newMedia->filename = $this->cleanFilename($file->getClientOriginalName());
+				$newMedia->type = $this->getTypeFromFile($file);
+
+				if (array_key_exists($index, $is_private) && ! empty($is_private[$index])) {
+					$uploadSuccess = $file->move(storage_path($newMedia->getDirectory()), $newMedia->filename);
+					$newMedia->private = true;
+				} else {
+					$uploadSuccess = $file->move(public_path($newMedia->getDirectory()), $newMedia->filename);
+				}
 
 				if (array_key_exists($index, $captions)) {
 					$newMedia->caption = $captions[$index];
 				}
 
-				$newMedia->filename = $filename;
-				$newMedia->type = $type;
-				$newMedia->save();
-
-				$data[] = [
-					'url' => action(__CLASS__ . '@edit', $newMedia->id),
-					'thumbnailUrl' => asset($newMedia->getFilename('library')),
-					'name' => $newMedia->caption ?: $newMedia->filename,
-					'deleteUrl' => action(__CLASS__ . '@destroy', $newMedia->id),
-					'deleteType' => 'DELETE'
-				];
+				if ($uploadSuccess) {
+					$newMedia->save();
+					$data[] = [
+						'url' => action(__CLASS__ . '@edit', $newMedia->id),
+						'thumbnailUrl' => asset($newMedia->getFilename('library')),
+						'name' => $newMedia->caption ?: $newMedia->filename,
+						'deleteUrl' => action(__CLASS__ . '@destroy', $newMedia->id),
+						'deleteType' => 'DELETE',
+						'id' => $newMedia->id,
+						'filename' => $newMedia->filename,
+						'type' => $newMedia->type,
+						'private' => array_key_exists($index, $is_private)
+					];
+				}
 			}
 		}
 
 		return Response::json(['files' => $data]);
+	}
+
+	/**
+	 * Return the sub-directory to save the uploaded file, based on the file's
+	 * mime type
+	 *
+	 * @param  Symfony\Component\HttpFoundation\File\UploadedFile  $file
+	 * @return string
+	 */
+	protected function getTypeFromFile(UploadedFile $file)
+	{
+		$mimeType = $file->getMimeType();
+
+		foreach($this->mimeTypeMapping as $regex => $directory) {
+			if (preg_match("#{$regex}#", $mimeType)) {
+				return $directory;
+			}
+		}
+
+		return 'misc';
 	}
 
 	/**