+++ src/migrations/2015_06_19_152123_add_private_to_media_table.php
@@ -1,0 +1,34 @@
+<?php
+
+use Illuminate\Database\Schema\Blueprint;
+use Illuminate\Database\Migrations\Migration;
+
+class AddPrivateToMediaTable extends Migration {
+
+	/**
+	 * Run the migrations.
+	 *
+	 * @return void
+	 */
+	public function up()
+	{
+		Schema::table('media', function(Blueprint $table)
+		{
+			$table->boolean('private');
+		});
+	}
+
+	/**
+	 * Reverse the migrations.
+	 *
+	 * @return void
+	 */
+	public function down()
+	{
+		Schema::table('media', function(Blueprint $table)
+		{
+			$table->dropColumn('private');
+		});
+	}
+
+}