+++ src/views/overview.blade.php
@@ -2,24 +2,31 @@
 
 @section('main')
 @parent
-	@include('admin::partials.new')
-	<h1>{{ $modelName }}</h1>
+	@section('report_header')
+		@include('admin::partials.new')
+		<h1>{{ $heading }}</h1>
 
-	@if (Session::has('model.updated'))
-		<div id="js-alert" class="alert alert-success" data-alert="alert">
-			{{ Session::get('model.updated') }}
-		</div>
-	@endif
+		@if (Session::has('model'))
+			@foreach(Session::get('model') as $msg)
+				<div id="js-alert" class="alert alert-success" data-alert="alert">
+					{{ $msg }}
+				</div>
+			@endforeach
+		@endif
 
-	@include('admin::partials.sort-alert')
+		@include('admin::partials.sort-alert')
 
-	{{ $report->getHeader() }}
+		{{ $report->getHeader() }}
+	@show
+	@section('report')
 	<div class="table-responsive">
 	@if ($report->hasRows())
-		<ol class="secret-list faux-table{{ $sortableClass }}" data-model="{{ $fullModelName }}">
+		<ol class="secret-list faux-table{{ $sortableClass }}" data-model="{{ $identifier }}">
 
 			<li class="faux-table-row faux-table-heading">
+			@if ($sortableClass)
 				<div class="faux-cell cell-small"></div>
+			@endif
 			@foreach ($report->getHeadings() as $heading)
 				<div class="faux-cell">{{ $heading }}</div>
 			@endforeach
@@ -28,9 +35,11 @@
 
 		@foreach ($report->getRows() as $row)
 			<li class="faux-table-row" data-id="{{ $row->getId() }}">
+			@if ($sortableClass)
 				<div class="faux-cell cell-small">
 					<i class="fa fa-sort sorting-handle"></i>
 				</div>
+			@endif
 			@foreach ($row->getColumns() as $name => $value)
 				<div class="faux-cell">{{ $value }}</div>
 			@endforeach
@@ -40,19 +49,25 @@
 						Edit
 					</a>
 
-					{{ Form::open(['class' => 'inline-form', 'action' => [ $controller . '@destroy', $row->getId() ], 'method' => 'DELETE']) }}
-						<button class="btn btn-danger btn-sm" data-warn="true" type="submit"><i class="fa fa-minus-square"></i> Delete</button>
-					{{ Form::close() }}
+					@if ($canDelete)
+						{{ Form::open(['class' => 'inline-form', 'action' => [ $controller . '@destroy', $row->getId() ], 'method' => 'DELETE']) }}
+							<button class="btn btn-danger btn-sm" data-warn="true" type="submit"><i class="fa fa-minus-square"></i> Delete</button>
+						{{ Form::close() }}
+					@endif
 				</div>
 			</li>
 		@endforeach
 		</ol>
-		{{ $report->getFooter() }}
 	@else
 		<p>Nothing here yet. Why not add something?</p>
 	@endif
 	</div>
-	@include('admin::partials.new')
+	@show
+
+	@section('report_footer')
+		{{ $report->getFooter() }}
+		@include('admin::partials.new')
+	@show
 
 	@section('scripts')
 		@parent