+++ src/views/partials/nav.blade.php
@@ -44,7 +44,7 @@
 			</ul>
 			<ul class="nav navbar-nav navbar-right">
 				<li class="dropdown">
-					<a href="#" class="dropdown-toggle" data-toggle="dropdown"><i class="fa fa-cog"></i>Logged in as {{ $user->name }} <b class="caret"></b></a>
+					<a href="#" class="dropdown-toggle" data-toggle="dropdown"><i class="fa fa-cog"></i>Logged in as {{ $user->first_name }} <b class="caret"></b></a>
 					<ul class="dropdown-menu">
 						<li><a href="/admin/users/{{ $user->id }}/edit"><i class="fa fa-wrench"></i> Edit Profile</a></li>
 						<li><a href="/"><i class="fa fa-desktop"></i> View Website</a></li>