+++ src/views/partials/nested-item.blade.php
@@ -1,14 +1,15 @@
 			<li id="nested-item_{{$row->getId()}}">
 				<div class="nested-group">
 					<div class="actions">
-						<a href="{{ URL::action($controller . '@edit', array($row->getModel()->id)) }}" class="btn btn-info btn-sm" type="submit">
+						<a href="{{ URL::action($controller . '@edit', array($row->getId())) }}" class="btn btn-info btn-sm" type="submit">
 							<i class="fa fa-pencil"></i>
 							Edit
 						</a>
-
-						{{ Form::model($row->getModel(), array('class' => 'inline-form', 'action' => array($controller . '@destroy', $row->getModel()->id), 'method' => 'DELETE')) }}
-							<button class="btn btn-danger btn-sm" data-warn="true" type="submit"><i class="fa fa-minus-square"></i> Delete</button>
-						{{ Form::close() }}
+						@if ($canDelete)
+							{{ Form::open(['class' => 'inline-form', 'action' => array($controller . '@destroy', $row->getId()), 'method' => 'DELETE']) }}
+								<button class="btn btn-danger btn-sm" data-warn="true" type="submit"><i class="fa fa-minus-square"></i> Delete</button>
+							{{ Form::close() }}
+						@endif
 					</div>
 					<div class="nested-value">
 						<i class="fa fa-sort sorting-handle"></i>